Remove laminas-cache configuration #minor
We installed `laminas/laminas-cache-storage-adapter-apcu` because it was a dependency of `doctrine/doctrine-module`. But if we bump the doctrine module, the dependency has been removed.

This is because doctrine-module ships with [Doctrine's own cache implementation](https://www.doctrine-project.org/projects/doctrine-cache/en/1.12/index.html) and the laminas-cache-* dependency was only for [sharing caches between Doctrine and Laminas](https://www.doctrine-project.org/projects/doctrine-module/en/5.0/caching.html#caching). We have no interest in doing that, so can drop the package completely.

We can also remove the apcu extension now that we're not using it