Fix dependabot pipeline (#49)
External PR builds don't have access to a write token (understandably), so Dependabot PRs cannot push their test results onto the PR comments.

To fix this, there is now a special "dependabot PR" build that only triggers for the dependabot user. When that is successfully completed, the main build job is started. Because it's now run from GitHub, not from an external user, it has the necessary permissions.

This solution was taken from [a comment on the Dependabot core repo](https://github.com/dependabot/dependabot-core/issues/3253#issuecomment-797125425).

#patch